Tweak the wording for helping build teams
Describing the process as a development consultant
on a team in a few bullet points is impossible.
We work with teams of all levels
from junior to expert
and are expected to solve problems of all shapes
and sizes.
All the while, we have to train that team
to pick up the mantle and be amazing
when we leave.